requirements.txt
# requirements.txt - BCI project
# python>=3.9,<3.12

# Core numeric / data
numpy>=1.24
pandas>=2.0
scipy>=1.10

# EEG/Signal processing
mne>=1.2
neurokit2>=0.2

# Streaming & hardware
pylsl>=1.17.0
pyserial>=3.5
brainflow>=4.8.0

# ML
scikit-learn>=1.2
torch>=2.0
torchvision>=0.15

# Utilities
matplotlib>=3.6
h5py>=3.8
joblib>=1.3

# Websocket & backend
fastapi>=0.95
uvicorn>=0.22
websockets>=11.0

# Dev / lint / packaging
black
pytest
python-dotenv

#
